backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
authlib
httpx
requests
itsdangerous
starlette